reporter: Move remaining "utils" classes to the top level
The top level already contains (more) utility-like classes, so
consolidate these classes there.
