+++ reporter/src/main/kotlin/ReportTableModel.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.reporter.utils
+package org.ossreviewtoolkit.reporter
 
 import java.util.SortedMap
 import java.util.SortedSet
